Use FlaUI.TestUtilities
