+++ src/FlaUI.Core.UITests/Elements/AutomationElementTests.cs
@@ -10,7 +10,7 @@ namespace FlaUI.Core.UITests.Elements
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.WinForms)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class AutomationElementTests : FlaUITestBase
+    public class AutomationElementTests : UITestBase
     {
         public AutomationElementTests(AutomationType automationType, TestApplicationType appType)
             : base(automationType, appType)
