+++ src/FlaUI.Core.UITests/Elements/CheckBoxTests.cs
@@ -9,7 +9,7 @@ namespace FlaUI.Core.UITests.Elements
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.WinForms)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class CheckBoxTests : FlaUITestBase
+    public class CheckBoxTests : UITestBase
     {
         public CheckBoxTests(AutomationType automationType, TestApplicationType appType)
             : base(automationType, appType)
