+++ src/FlaUI.Core.UITests/Elements/ComboBoxTests.cs
@@ -11,7 +11,7 @@ namespace FlaUI.Core.UITests.Elements
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.WinForms)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class ComboBoxTests : FlaUITestBase
+    public class ComboBoxTests : UITestBase
     {
         private Window _mainWindow;
