+++ src/FlaUI.Core.UITests/Elements/GridTests.cs
@@ -9,7 +9,7 @@ namespace FlaUI.Core.UITests.Elements
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.WinForms)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class GridTests : FlaUITestBase
+    public class GridTests : UITestBase
     {
         private Grid _grid;
