+++ src/FlaUI.Core.UITests/Elements/LabelTests.cs
@@ -8,7 +8,7 @@ namespace FlaUI.Core.UITests.Elements
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.WinForms)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class LabelTests : FlaUITestBase
+    public class LabelTests : UITestBase
     {
         public LabelTests(AutomationType automationType, TestApplicationType appType)
             : base(automationType, appType)
