+++ src/FlaUI.Core.UITests/Elements/MenuTests.cs
@@ -9,7 +9,7 @@ namespace FlaUI.Core.UITests.Elements
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.WinForms)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class MenuTests : FlaUITestBase
+    public class MenuTests : UITestBase
     {
         public MenuTests(AutomationType automationType, TestApplicationType appType)
             : base(automationType, appType)
