+++ src/FlaUI.Core.UITests/Elements/PopupTests.cs
@@ -7,7 +7,7 @@ namespace FlaUI.Core.UITests.Elements
 {
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class PopupTests : FlaUITestBase
+    public class PopupTests : UITestBase
     {
         public PopupTests(AutomationType automationType, TestApplicationType appType)
             : base(automationType, appType)
