+++ src/FlaUI.Core.UITests/Elements/TableTests.cs
@@ -9,7 +9,7 @@ namespace FlaUI.Core.UITests.Elements
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.WinForms)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class TableTests : FlaUITestBase
+    public class TableTests : UITestBase
     {
         private Grid _table;
