+++ src/FlaUI.Core.UITests/Elements/TextBoxTests.cs
@@ -11,7 +11,7 @@ namespace FlaUI.Core.UITests.Elements
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.WinForms)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class TextBoxTests : FlaUITestBase
+    public class TextBoxTests : UITestBase
     {
         private const string DefaultTextBoxText = "Test TextBox";
