+++ src/FlaUI.Core.UITests/Elements/WindowTests.cs
@@ -9,7 +9,7 @@ namespace FlaUI.Core.UITests.Elements
     [TestFixture(AutomationType.UIA2, TestApplicationType.Wpf)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.WinForms)]
     [TestFixture(AutomationType.UIA3, TestApplicationType.Wpf)]
-    public class WindowTests : FlaUITestBase
+    public class WindowTests : UITestBase
     {
         public WindowTests(AutomationType automationType, TestApplicationType appType)
             : base(automationType, appType)
