+++ src/FlaUI.Core.UITests/FlaUI.Core.UITests.csproj
@@ -9,7 +9,7 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>FlaUI.Core.UITests</RootNamespace>
     <AssemblyName>FlaUI.Core.UITests</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
   </PropertyGroup>
@@ -31,17 +31,27 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="nunit.framework, Version=3.8.1.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
-      <HintPath>..\packages\NUnit.3.8.1\lib\net45\nunit.framework.dll</HintPath>
+    <Reference Include="Accessibility" />
+    <Reference Include="FlaUI.TestUtilities, Version=0.1.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\FlaUI.TestUtilities.0.1.0\lib\netstandard2.0\FlaUI.TestUtilities.dll</HintPath>
     </Reference>
+    <Reference Include="nunit.framework, Version=3.12.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
+      <HintPath>..\packages\NUnit.3.12.0\lib\net45\nunit.framework.dll</HintPath>
+    </Reference>
+    <Reference Include="PresentationCore" />
+    <Reference Include="PresentationFramework" />
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Drawing" />
+    <Reference Include="System.IO.Compression" />
+    <Reference Include="System.Management" />
+    <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Data" />
     <Reference Include="System.Xml" />
+    <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ApplicationTests.cs" />
@@ -87,11 +97,9 @@
     <Compile Include="Patterns\ScrollItemPatternTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SearchTests.cs" />
-    <Compile Include="TestFramework\FlaUITestBase.cs" />
-    <Compile Include="TestFramework\TestApplicationType.cs" />
-    <Compile Include="TestFramework\TestUtilities.cs" />
     <Compile Include="TestFramework\UITestBase.cs" />
-    <Compile Include="TestFramework\NUnitProgressLogger.cs" />
+    <Compile Include="TestFramework\TestApplicationType.cs" />
+    <Compile Include="TestFramework\UtilityMethods.cs" />
     <Compile Include="TouchTests.cs" />
     <Compile Include="WordPad\Screens\InfoScreen.cs" />
     <Compile Include="WordPad\Screens\MainScreen.cs" />
@@ -100,7 +108,7 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\FlaUI.Core\FlaUI.Core.csproj">
-      <Project>{4d98a40f-e35d-4a88-8428-8462af75a4cf}</Project>
+      <Project>{df81078d-0932-47ba-b5a7-176cff202375}</Project>
       <Name>FlaUI.Core</Name>
     </ProjectReference>
     <ProjectReference Include="..\FlaUI.UIA2\FlaUI.UIA2.csproj">
