+++ src/FlaUI.Core.UITests/packages.config
@@ -1,4 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NUnit" version="3.8.1" targetFramework="net45" />
+  <package id="FlaUI.TestUtilities" version="0.1.0" targetFramework="net48" />
+  <package id="NUnit" version="3.12.0" targetFramework="net45" />
+  <package id="System.Management" version="4.7.0" targetFramework="net48" />
 </packages>
