+++ src/FlaUI.Core.UnitTests/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Castle.Core" version="4.4.0" targetFramework="net45" />
-  <package id="NSubstitute" version="4.2.1" targetFramework="net45" />
+  <package id="NSubstitute" version="4.2.1" targetFramework="net45" requireReinstallation="true" />
   <package id="NUnit" version="3.8.1" targetFramework="net45" />
   <package id="System.Threading.Tasks.Extensions" version="4.3.0" targetFramework="net45" />
 </packages>
